// src/model/ts/interface/producto.ts
export interface ProductoFilterInterface {
  categoria: string
  codigo: string
  codigoBarras: string
  costo: number
  idMedida: string
  idProducto: string
  idTipoProducto: string
  idTipoTratamientoProducto: string
  idMetodoDepreciacion: string
  imagen: string | null
  nombre: string
  precio: number
  sku: string
  tipoProducto: string
  unidad: string
}